+++ test/DriverFactoryTest.php
@@ -4,12 +4,13 @@ declare(strict_types=1);
 
 namespace RoaveTest\PsrContainerDoctrine;
 
-use Doctrine\Common\Persistence\Mapping\Driver\MappingDriverChain;
 use Doctrine\ORM\Mapping\Driver;
-use OutOfBoundsException;
+use Doctrine\Persistence\Mapping\Driver\FileDriver;
+use Doctrine\Persistence\Mapping\Driver\MappingDriverChain;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\ContainerInterface;
 use Roave\PsrContainerDoctrine\DriverFactory;
+use Roave\PsrContainerDoctrine\Exception\OutOfBoundsException;
 
 class DriverFactoryTest extends TestCase
 {
@@ -44,10 +45,12 @@ class DriverFactoryTest extends TestCase
         $factory = new DriverFactory();
 
         $driver = $factory($container->reveal());
+        $this->assertInstanceOf(FileDriver::class, $driver);
         $this->assertSame($globalBasename, $driver->getGlobalBasename());
     }
 
     /**
+     * @psalm-param class-string<FileDriver> $driverClass
      * @dataProvider simplifiedDriverClassProvider
      */
     public function testItSupportsSettingExtensionInDriversUsingSymfonyFileLocator(string $driverClass) : void
@@ -67,14 +70,16 @@ class DriverFactoryTest extends TestCase
             ],
         ]);
 
-        $factory = new DriverFactory();
-
-        $driver = $factory($container->reveal());
-        $this->assertInstanceOf($driverClass, $driver);
+        $driver = (new DriverFactory())->__invoke($container->reveal());
+        $this->assertInstanceOf(FileDriver::class, $driver);
         $this->assertSame($extension, $driver->getLocator()->getFileExtension());
     }
 
-    /** @return string[][] */
+    /**
+     * @return string[][]
+     *
+     * @psalm-return list<list<class-string<FileDriver>>>
+     */
     public function simplifiedDriverClassProvider() : array
     {
         return [
@@ -104,6 +109,7 @@ class DriverFactoryTest extends TestCase
         $factory = new DriverFactory();
 
         $driver = $factory($container->reveal());
+        $this->assertInstanceOf(MappingDriverChain::class, $driver);
         $this->assertInstanceOf(TestAsset\StubFileDriver::class, $driver->getDefaultDriver());
     }
 }