+++ test/TestAsset/StubFileDriver.php
@@ -4,22 +4,23 @@ declare(strict_types=1);
 
 namespace RoaveTest\PsrContainerDoctrine\TestAsset;
 
-use Doctrine\Common\Persistence\Mapping\Driver\FileDriver;
 use Doctrine\Persistence\Mapping\ClassMetadata;
+use Doctrine\Persistence\Mapping\Driver\FileDriver;
 
 class StubFileDriver extends FileDriver
 {
-    // Disable these sniffs, since we can't control inheritance
-    // phpcs:disable SlevomatCodingStandard.TypeHints.ReturnTypeHint.MissingAnyTypeHint
-    // phpcs:disable SlevomatCodingStandard.TypeHints.ParameterTypeHint.MissingAnyTypeHint
-    protected function loadMappingFile($file)
+    /**
+     * {@inheritdoc}
+     */
+    protected function loadMappingFile($file) : array
     {
         return [];
     }
 
+    /**
+     * @param string|class-string $className
+     */
     public function loadMetadataForClass($className, ClassMetadata $metadata) : void
     {
     }
-
-    // phpcs:enable
 }